Remove some unnecessary things. Create pending repo for packages that should be merged upstream.
