+++ configs/base/repos.yaml
@@ -1,2 +1,3 @@
 repos:
   - $spack/../repos/exawind
+  - $spack/../repos/pending
