--- repos/exawind/packages/hypre/ibm-ppc64le.patch
@@ -1,12 +1,0 @@
---- hypre-2.11.1/src/config/config.guess	2016-06-09 10:56:18.000000000 -0400
-+++ hypre-2.11.1/src/config/config.guess.patch	2017-02-24 22:53:42.718337013 -0500
-@@ -882,6 +882,9 @@
-     ppc64:Linux:*:*)
- 	echo powerpc64-unknown-linux-gnu
- 	exit ;;
-+    ppc64le:Linux:*:*)
-+	echo powerpc64le-unknown-linux-gnu
-+	exit ;;
-     alpha:Linux:*:*)
- 	case `sed -n '/^cpu model/s/^.*: \(.*\)/\1/p' < /proc/cpuinfo` in
- 	  EV5)   UNAME_MACHINE=alphaev5 ;;
