--- repos/exawind/packages/hypre/superlu-dist-link-2.15.0.patch
@@ -1,13 +1,0 @@
-diff --git a/src/lib/Makefile b/src/lib/Makefile
-index f49d1683..615947b0 100644
---- a/src/lib/Makefile
-+++ b/src/lib/Makefile
-@@ -58,7 +58,7 @@ $(BLASFILES)\
- $(LAPACKFILES)
- 
- SONAME = libHYPRE-${HYPRE_RELEASE_VERSION}${HYPRE_LIB_SUFFIX}
--SOLIBS = ${MPILIBDIRS} ${MPILIBS} ${LAPACKLIBDIRS} ${LAPACKLIBS}\
-+SOLIBS = ${DSUPERLU_LIBS} ${MPILIBDIRS} ${MPILIBS} ${LAPACKLIBDIRS} ${LAPACKLIBS}\
-  ${BLASLIBDIRS} ${BLASLIBS} ${LIBS} ${FLIBS}
- 
-
