+++ repos/exawind/packages/trilinos/package.py
@@ -1,6 +1,4 @@
-from spack import *
 from spack.pkg.builtin.trilinos import Trilinos as bTrilinos
-import os
 
 class Trilinos(bTrilinos):
     depends_on('ninja', type='build')
