+++ repos/pending/packages/hypre/package.py
@@ -1,0 +1,4 @@
+from spack.pkg.builtin.hypre import Hypre as bHypre
+
+class Hypre(bHypre):
+    variant('shared', default=True, description="Build shared library (disables static library)")
