+++ repos/pending/repo.yaml
@@ -1,0 +1,2 @@
+repo:
+  namespace: 'pending'
