--- scripts/test-scripts/com.update.build.test.repo.nightly.plist
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-  <key>Label</key>
-  <string>com.update.build.test.repo.nightly</string>
-
-  <key>ProgramArguments</key>
-  <array>
-    <string>/bin/bash</string>
-    <string>-c</string>
-    <string>cd /Users/jrood/nalu-wind-testing/build-test && ./test-scripts/update-build-test-repo.sh</string>
-  </array>
-
-  <key>UserName</key>
-  <string>jrood</string>
-
-  <key>Nice</key>
-  <integer>1</integer>
-
-  <key>StartCalendarInterval</key>
-  <dict>
-      <key>Hour</key>
-      <integer>0</integer>
-      <key>Minute</key>
-      <integer>0</integer>
-  </dict>
-
-  <key>StandardErrorPath</key>
-  <string>/Users/jrood/nalu-wind-testing/logs/last-build-test-repo-update.err</string>
-
-  <key>StandardOutPath</key>
-  <string>/Users/jrood/nalu-wind-testing/logs/last-build-test-repo-update.out</string>
-</dict>
-</plist>
